+++ src/hooks/useTypeaheadQuery.ts
@@ -7,28 +7,23 @@ import {
   useState,
 } from 'react';
 
-import { GenericDataType } from '../common';
 import { useDebouncedValue } from './useDebouncedValue';
 
-export interface UseTypeaheadQueryOptions<DataItem> {
+export interface UseTypeaheadQueryOptions {
   debounceTime?: number;
   onDebouncedChange?: (value: string) => void;
-  options?: DataItem[];
 }
 
 export interface UseTypeaheadQueryResult {
-  isLoading: boolean;
   query: string;
   setQuery: Dispatch<SetStateAction<string>>;
 }
 
-export const useTypeaheadQuery = <DataItem extends GenericDataType>({
+export const useTypeaheadQuery = ({
   debounceTime,
   onDebouncedChange,
-  options,
-}: UseTypeaheadQueryOptions<DataItem>): UseTypeaheadQueryResult => {
+}: UseTypeaheadQueryOptions): UseTypeaheadQueryResult => {
   const onDebouncedChangeFn = useRef(onDebouncedChange);
-  const [isLoading, setIsLoading] = useState(false);
   const [query, setQuery] = useState('');
   const { debouncedValue } = useDebouncedValue<string>(
     query,
@@ -41,14 +36,7 @@ export const useTypeaheadQuery = <DataItem extends GenericDataType>({
   useEffect(() => {
     onDebouncedChangeFn.current?.(currentQuery);
   }, [currentQuery]);
-  useEffect(() => {
-    if (options !== undefined) setIsLoading(false);
-  }, [options]);
-  useEffect(() => {
-    setIsLoading(query !== '');
-  }, [query]);
   return {
-    isLoading,
     query,
     setQuery,
   };